Moved fuzz_tests to _fuzz_testing as it is internally used.
