+++ python/cudf/cudf/_fuzz_testing/tests/readme.md
@@ -39,5 +39,5 @@ python write_csv.py csv_writer_test
 
 Using the `pythonfuzz` decorator pass in `regression=True` with `dirs` having list of directories 
 ```python
-@pythonfuzz(data_handle=CSVWriter, regression=True, dir=["/cudf/python/cudf/cudf/fuzz_tests"])
+@pythonfuzz(data_handle=CSVWriter, regression=True, dir=["/cudf/python/cudf/cudf/_fuzz_testing"])
 ```
